+++ catalog/app/containers/Bucket/PackageCreateDialog.js
@@ -76,8 +76,8 @@ const useFilesInputStyles = M.makeStyles((t) => ({
   active: {
     background: t.palette.action.selected,
   },
-  draggable: {
-    outline: `2px dashed ${t.palette.primary.main}`,
+  outlined: {
+    outline: `2px dashed ${t.palette.primary.light}`,
     outlineOffset: '-2px',
   },
   dropMsg: {
@@ -275,7 +275,7 @@ function FilesInput({
               isDragActive && !disabled && classes.active,
               !!error && classes.dropzoneErr,
               !error && warn && classes.dropzoneWarn,
-              isDragging && classes.draggable,
+              isDragging && classes.outlined,
             ),
           })}
         >
@@ -570,21 +570,28 @@ function PackageCreateDialog({
 
   const [editorElement, setEditorElement] = React.useState()
 
+  const resizeObserver = React.useMemo(
+    () =>
+      new window.ResizeObserver((entries) => {
+        const { height } = entries[0]?.contentRect
+        setMetaHeight(height)
+      }),
+    [setMetaHeight],
+  )
+
   const onFormChange = React.useCallback(
     ({ dirtyFields, values }) => {
-      if (document.body.contains(editorElement)) {
-        setMetaHeight(editorElement.clientHeight)
-      }
       if (dirtyFields.name) handleNameChange(values.name)
     },
-    [editorElement, handleNameChange, setMetaHeight],
+    [handleNameChange],
   )
 
   React.useEffect(() => {
-    if (document.body.contains(editorElement)) {
-      setMetaHeight(editorElement.clientHeight)
+    if (editorElement) resizeObserver.observe(editorElement)
+    return () => {
+      if (editorElement) resizeObserver.unobserve(editorElement)
     }
-  }, [editorElement, setMetaHeight])
+  }, [editorElement, resizeObserver])
 
   const username = redux.useSelector(authSelectors.username)
   const usernamePrefix = React.useMemo(() => PD.getUsernamePrefix(username), [username])